[#3123] 조회 api 응답 상태 값 수정

+++ Billing.Data/Interfaces/IServiceHierachyRepository.cs
@@ -6,7 +6,7 @@ namespace Billing.Data.Interfaces
     {
         Task<ServiceHierarchy> Add(ServiceHierarchy entity);
         Task Delete(long serialNo);
-        Task<ServiceHierarchy?> Get(int serialNo);
+        Task<ServiceHierarchy?> Get(long serialNo);
         Task<List<ServiceHierarchy>> GetChild(long parentAccountId);
         Task<ServiceHierarchy?> GetParent(long accountId);
         Task Update(ServiceHierarchy entity);
